XR Session support check & session event
Now that an XR device has been obtained, we need to check if it can support an 'immersive' session (Not sure if needed for magic window environments). If an immersive session can be supported, head to the xrInit function to setup class fields and later request a session from the XR api. Also gives the webgl renderer's vr context the device and context.