Add animator file.
Handles animations for all the scenes. Not working right now. Problem seems to be that the old animation frame loops arent ending when a new scene is instantiated. Class update functions calls are replicated and compounded to the point where clicking a link significantly kills the site performance. Need to fix.