Magic Window Added
Used WebXR API to check browser/device compatibility with immersive and non-immersive sessions.
xrController.js is run first to validate and setup an XR Session object
with a reference frame and a canvas to render to. Provides all the
necessary information to be able to interface with the THREE.JS
renderer.

XRScene now recieves the XRSession information for obtaining and
utilizing XR transformations in the THREE.JS camera and renderer. If a
session has not been created, the animator proceeds to render the scene
normally. Otherwise the scene, camera, and renderer matrices are morphed
using the device pose returned from the XRSession's frames.