Fix incorrect magic window canvas sizing
The magic window canvas was not inheriting the width and height of the window on its creation so its size was incorrect and its projection was bloated.  Also updated the magic window rendering viewport to get its height and width from the magic window canvas. Finally, updated the resize function so that it resized the current vr-port canvas being displayed by running an id search every resize.