Fix renderer resizing
Two things were preventing two eyes to render correctly in this file. The first was that we were updating the renderer viewport through a three.js function called setViewport. This is incorrect and we should be using the webgl context function viewport(). Second is that the renderer wasn't being resized to the xrLayer framebuffer dimensions. Changing these two errors fixed the distortions that the two eye rendering was showing.