Call animationFrame from xrController
Set up an event listener in xrController that calls a function in xrScene that resets the animation loop accordingly. Needed to do this because each new XR session cancels the previous animation loops. Also canceling the last loop just in case the user is in the initial state of the site in which case there would be two loops running at once. So fixed that.