Controller class
Added a controller class to manage and update XR input controller models. Holds a list of the models that are currently in use. Adds controller models based on their type. Removes controllers based on index from list using a splice. Updates a controller based on its index and a transform matrix passed in.