Render controller into scene
Controller is rendered into the scene for each frame. Currently, only local transforms are taking place and no raycasting is active. If the number of controllers for each scene is less than the number of input sources then add a controller to the controllers list. Then for that controller, apply the input pose transformations to it.

I also updated the translate matrix to just return the translation matrix rather than do the premultiply itself. Makes the function more modular.