Controller position correctly updates with user position
Before, the userPosition field was an inverted position due to the calculations done in updateTouchPosition. This is a confusing value for something called user position as it actually denotes the transformation that needs to be applied to the world origin rather than the objects within. I changed this so that the value was no longer inverted and then added a new function translateObjectMatrix that updates the position of the object matrix passed in by the offset position given. The controller position now places itself at the user position. Next up, rotations.