Calculate raycaster position and direction
The raycaster origin and direction vectors are being calculated from the inputPose.targetRay matrix and are being passed into the raycaster class. After the raycaster is updated, an intersection is obtained from a list of TriggerMesh objects to call onTrigger events for based on the type of intersection event. The laser length is also updated such that it is the distance from the controller to the intersection.