Handle select events
Created event listeners for selectstart and selectend that fire events inside of the xrScene class which handle button pressed selection of objects. Also added an example in the home scene with two selectable objects. On hover, an item is colored red. On selection hold an item is colored green. On selection release the item is colored blue before on hover coloring is resumed from a click elsewhere or cursor moving off of object.