Multiply intersection.point by inverse scene world matrix
The intersection point provided by the raycaster was from transformed world coordinates, meaning that while correct in the world, it was incorrect for scene local coordinates. Since our triggerMeshes are in scene local coordinates, during XR sessions we need to multiply the intersection.point by the inverse of the scene world matrix to negate the effects of the world transformation. This gives us localized scene coordinates for every intersection and provides the correct intersection information for every triggerMesh event. I also made sure to check for keyboard controls enabled so that the inverse multiplication wouldn't take place in regular raycasting scenarios.