Using Box3s to test boundaries.
Scene boundaries are now set through a series of bounding boxes that the user must remain within. When the user exits a bounding box, the closest bounding box to the user is found and their position is clamped to it. If the user teleports out of the box, they are clamped to the closest bounding box to the teleportation point. User can translate into intersecting bounding boxes. It is also possible to set the bounding box from an Object3D as shown in the Home.js example.